# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GoCreate API Help Page",
    "description": "**This API exposes core commerce data for B2B clients, providing visibility into products, pricing, inventory, and operational status. It is optimized for client-facing applications that require secure, reliable, and consistent data access.**\r\n\r\n**Data Retention Policy:** Starting January 14, 2026, we will provide order details for ongoing orders. After an order is delivered, its details will remain available only for 4 weeks.",
    "contact": {
      "name": "GoCreate Support Team",
      "url": "https://api.gocreate.nu"
    },
    "version": "2.0"
  },
  "paths": {
...
    "/Fabric/Post": {
      "post": {
        "tags": [
          "Fabric"
        ],
        "summary": "Post",
        "description": "This endpoint validates the incoming request parameters and returns accurate, structured fabric information.",
        "operationId": "FabricPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FabricStockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FabricStockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseApiResponse"
                },
                "example": {
                  "IsValidResult": false,
                  "ErrorCode": "400",
                  "ErrorMessage": "Validation failed.",
                  "ID": 0
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseApiResponse"
                },
                "example": {
                  "IsValidResult": false,
                  "ErrorCode": "401",
                  "ErrorMessage": "Authentication failed.",
                  "ID": 0
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseApiResponse"
                },
                "example": {
                  "IsValidResult": false,
                  "ErrorCode": "403",
                  "ErrorMessage": "Access denied.",
                  "ID": 0
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseApiResponse"
                },
                "example": {
                  "IsValidResult": false,
                  "ErrorCode": "404",
                  "ErrorMessage": "Resource not found.",
                  "ID": 0
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseApiResponse"
                },
                "example": {
                  "IsValidResult": false,
                  "ErrorCode": "500",
                  "ErrorMessage": "Internal server error.",
                  "ID": 0
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseApiResponse": {
        "type": "object",
        "properties": {
          "IsValidResult": {
            "type": "boolean",
            "description": "Indicates whether the result is valid."
          },
          "ErrorCode": {
            "type": "string",
            "description": "The error code when the response is not valid.",
            "nullable": true
          },
          "ErrorMessage": {
            "type": "string",
            "description": "A human-readable error message when the response is not valid.",
            "nullable": true
          },
          "Id": {
            "type": "integer",
            "description": "The identifier associated with this response.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "FabricStockInfoAPIResult": {
        "type": "object",
        "properties": {
          "Code": {
            "type": "string",
            "description": "The fabric code.",
            "nullable": true
          },
          "Bunch": {
            "type": "string",
            "description": "The fabric bunch.",
            "nullable": true
          },
          "Collection": {
            "type": "string",
            "description": "The fabric collection.",
            "nullable": true
          },
          "ImageUrl": {
            "type": "string",
            "description": "URL to the fabric image.",
            "nullable": true
          },
          "Supplier": {
            "type": "string",
            "description": "The fabric supplier.",
            "nullable": true
          },
          "Season": {
            "type": "string",
            "description": "The season.",
            "nullable": true
          },
          "RequestedId": {
            "type": "integer",
            "description": "The requested identifier.",
            "format": "int32"
          },
          "RequestedCode": {
            "type": "string",
            "description": "The requested code.",
            "nullable": true
          },
          "AtelierName": {
            "type": "string",
            "description": "The atelier name.",
            "nullable": true
          },
          "Id": {
            "type": "integer",
            "description": "The internal identifier.",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "description": "The fabric name.",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "description": "The fabric description.",
            "nullable": true
          },
          "PPriceCategories": {
            "type": "string",
            "description": "The price categories.",
            "nullable": true
          },
          "CutLengthFabric": {
            "type": "boolean",
            "description": "Whether the fabric is cut length."
          },
          "Stock": {
            "type": "number",
            "description": "The available stock quantity.",
            "format": "double"
          },
          "ExtraDays": {
            "type": "integer",
            "description": "The Fabric Extra days",
            "format": "int32"
          },
          "RPrice": {
            "type": "number",
            "description": "The retail price.",
            "format": "double"
          },
          "SoldOutSince": {
            "type": "string",
            "description": "Date when the fabric was marked sold out.",
            "format": "date-time",
            "nullable": true
          },
          "Status": {
            "type": "string",
            "description": "The fabric status.",
            "nullable": true
          },
          "StatusId": {
            "type": "integer",
            "description": "The fabric status identifier.",
            "format": "int32"
          },
          "RequestedName": {
            "type": "string",
            "description": "The requested fabric name.",
            "nullable": true
          },
          "MessageCode": {
            "type": "string",
            "description": "The message code.",
            "nullable": true
          },
          "Message": {
            "type": "string",
            "description": "The message text.",
            "nullable": true
          },
          "UsedFor": {
            "type": "string",
            "description": "What the fabric is used for.",
            "nullable": true
          },
          "Composition": {
            "type": "string",
            "description": "The fabric composition.",
            "nullable": true
          },
          "DurabilityIndicator": {
            "type": "string",
            "description": "The durability indicator.",
            "nullable": true
          },
          "FabricOnOrder": {
            "type": "number",
            "description": "The quantity of fabric currently on order.",
            "format": "double"
          },
          "FabricReadyDate": {
            "type": "string",
            "description": "The date when fabric will be ready.",
            "format": "date-time",
            "nullable": true
          },
          "IsCustomerOwn": {
            "type": "boolean",
            "description": "Indicates whether the fabric is customer-owned."
          }
        },
        "additionalProperties": false
      },
      "FabricStockRequest": {
        "required": [
          "AuthenticationToken",
          "FabricInputs",
          "Password",
          "UserName"
        ],
        "type": "object",
        "properties": {
          "UserName": {
            "minLength": 1,
            "type": "string",
            "description": "Username for authentication"
          },
          "Password": {
            "minLength": 1,
            "type": "string",
            "description": "Password for authentication"
          },
          "AuthenticationToken": {
            "minLength": 1,
            "type": "string",
            "description": "Authentication token for authorized requests"
          },
          "PageNo": {
            "type": "integer",
            "description": "The page number.",
            "format": "int32",
            "nullable": true
          },
          "PageSize": {
            "type": "integer",
            "description": "The size of the page.",
            "format": "int32",
            "nullable": true
          },
          "FabricInputs": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MaterialInput"
            },
            "description": "The fabric inputs."
          }
        },
        "additionalProperties": false
      },
      "FabricStockResponse": {
        "type": "object",
        "properties": {
          "IsValidResult": {
            "type": "boolean",
            "description": "Indicates whether the result is valid."
          },
          "ErrorCode": {
            "type": "string",
            "description": "The error code when the response is not valid.",
            "nullable": true
          },
          "ErrorMessage": {
            "type": "string",
            "description": "A human-readable error message when the response is not valid.",
            "nullable": true
          },
          "Id": {
            "type": "integer",
            "description": "The identifier associated with this response.",
            "format": "int32"
          },
          "FabricStockResult": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FabricStockInfoAPIResult"
            },
            "description": "The fabric stock result.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MaterialInput": {
        "required": [
          "ID"
        ],
        "type": "object",
        "properties": {
          "ID": {
            "type": "integer",
            "description": "The material identifier.",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "description": "The material name.",
            "nullable": true
          },
          "Code": {
            "type": "string",
            "description": "The material code.",
            "nullable": true
          },
          "PriceCategoryId": {
            "type": "integer",
            "description": "The price category identifier.",
            "format": "int32",
            "nullable": true
          },
          "Status": {
            "type": "integer",
            "description": "The material status identifier.",
            "format": "int32",
            "nullable": true
          },
          "XtraDays": {
            "type": "integer",
            "description": "Extra days (lead time) for the material.",
            "format": "int32",
            "nullable": true
          },
          "SearchStartWith": {
            "type": "boolean",
            "description": "Whether the search should match the start of the name."
          },
          "CombinationId": {
            "type": "integer",
            "description": "The combination identifier.",
            "format": "int32"
          },
          "AtelierIds": {
            "type": "string",
            "description": "Comma-separated atelier identifiers.",
            "nullable": true
          },
          "IsCLFabricSearch": {
            "type": "integer",
            "description": "Indicator whether this is a cut-length fabric search (nullable).",
            "format": "int32",
            "nullable": true
          },
          "ProductLineId": {
            "type": "integer",
            "description": "The product line identifier.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}